(extron:sw_usb) add a test module
Sometimes the host does not detect the switch or the device hangs when
switching host. This switches to 0 (which may not do anything at least
according to the manual) before switching to the correct input after
half a second. We hope that half second gives devices and hosts a
little more time to manage the change in state.